% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnetworks.R
\name{run_degree_preserving_randomization}
\alias{run_degree_preserving_randomization}
\title{Perform global degree-preserving randomization and compute network metrics.}
\usage{
run_degree_preserving_randomization(
  graph,
  smoops_nodes,
  num_randomizations = 1000,
  niter = 1000,
  seed = 42
)
}
\arguments{
\item{graph}{An igraph object representing the full network.}

\item{smoops_nodes}{A character vector of node names corresponding to the smOOPs subgraph.}

\item{num_randomizations}{Integer, the number of randomizations to perform (default is 1000).}

\item{niter}{Integer, the number of iterations for the degree-preserving rewiring algorithm (default is 1000).}

\item{seed}{Integer, the random seed for reproducibility (default is 42).}
}
\value{
A list containing two elements:
\itemize{
\item \code{random_metrics}: A list of randomization metrics, including \code{lcc_sizes}, \code{avg_degrees}, \code{edge_densities}, and \code{clustering_coeffs}.
\item \code{observed_metrics}: A list of observed metrics for the original smOOPs subgraph, including \code{lcc_size}, \code{avg_degree}, \code{edge_density}, and \code{clustering_coeff}.
}
}
\description{
This function performs a degree-preserving randomization of the input graph by rewiring edges while maintaining the degree distribution.
It then compares various network metrics (largest connected component size, average degree, edge density, and global clustering coefficient)
between the observed smOOPs subgraph and the subgraphs generated from the randomized networks. The results of the randomizations and the observed metrics are returned.
}
